Create the class in "main" scope, moving from test.
